save landmarks to binary and transform Weight to u32

+++ hermes_core/src/constants.rs
@@ -2,6 +2,6 @@ use crate::weighting::Weight;
 
 pub(crate) const INVALID_NODE: usize = usize::MAX;
 pub(crate) const INVALID_EDGE: usize = usize::MAX;
-pub(crate) const MAX_WEIGHT: Weight = usize::MAX;
+pub(crate) const MAX_WEIGHT: Weight = u32::MAX;
 
 pub(crate) const EARTH_RADIUS_METERS: f64 = 6_371_000.0;
